Fixed downloading invoice with DinkPdf and MySql seed data #725 (#728)
* Added AspnetCore directory helper see https://github.com/aspnet/AspNetCore/issues/4206#issuecomment-445612167. Sets current directory in `SimplCommerce.WebHost.Program.cs` because we're unable to use `Directory.GetCurrentDirectory`

* Fixed: DinkPdf download invoice `Unable to load DLL 'libwkhtmltox' or one of its dependencies:`. Added CustomAssemblyLoadContext and load the dll in DinkPdf ConfigureServices.

* Fixed MySql seed data. Color Key/ Display arrays missing quotes.

* Fixed payment process not going through and showing last step. Changed amount to `decimal` and removed casting to int.

* Fixed: MySql Fashion product seed data bad datetime format. CreatedOn, UpdatedOn

* Fixed: Clicking `Do it!` to reset `MySql` SampleData to Phones or Fashion. Deletes the wishlist table.

* Made Stripe checkout Amount long for zero decimal. #726

+++ src/Modules/SimplCommerce.Module.PaymentStripe/Areas/PaymentStripe/Components/StripeLandingViewComponent.cs
@@ -42,7 +42,7 @@ namespace SimplCommerce.Module.PaymentStripe.Areas.PaymentStripe.Components
             var regionInfo = new RegionInfo(CultureInfo.CurrentCulture.LCID);
             var model = new StripeCheckoutForm();
             model.PublicKey = stripeSetting.PublicKey;
-            model.Amount = (int)zeroDecimalAmount;
+            model.Amount = (long)zeroDecimalAmount;
             model.ISOCurrencyCode = regionInfo.ISOCurrencySymbol;
 
             return View(this.GetViewPath(), model);
